backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
numpy==1.26.4
scipy==1.12.0
pandas==2.2.0
pydantic==2.6.1
requests==2.31.0
arch==6.3.0
hmmlearn==0.3.0
torch==2.2.0
scikit-learn==1.4.0
httpx==0.27.0
pytest==8.0.0
pytest-asyncio==0.23.4
python-dotenv==1.0.1
